+++ src/binarytodecimal.cpp
