+++ src/hcf.cpp
