+++ src/modularMultiplicativeInverse.cpp
